Split the event and worker blocks into eight extension files
Verbatim moves out of AppModel.swift into AppModel+SettingsAndProvider,
+OverviewActions, +InspectorLayout, +PermissionsAndCapabilities,
+ChatWorkers, +SessionActivity, +BackendEvents, and +WorkspaceProfiles.
Private members the moved code still needs are widened to internal with
breadcrumb comments; behavior-identical.
